fix: remove components with async await from storybook
