+++ src/stories/components/tab.stories.ts
@@ -1,13 +1,36 @@
-import type { Meta, StoryObj } from "@storybook/react";
-import { Tab } from "@/components/tab";
+import type { Meta, StoryObj } from '@storybook/react'
+import { expect, userEvent, within } from '@storybook/test'
+import { Tab } from '@/components/tab'
 
 const meta = {
-  title: "Components/Tab",
+  title: 'Components/Tab',
   component: Tab,
-} satisfies Meta<typeof Tab>;
+} satisfies Meta<typeof Tab>
 
-export default meta;
+export default meta
 
-type Story = StoryObj<typeof meta>;
+type Story = StoryObj<typeof meta>
 
-export const TabComponent: Story = {};
+export const TabComponent: Story = {
+  play: async ({ canvasElement }) => {
+    const canvas = within(canvasElement)
+
+    const listButton = canvas.getByLabelText(/list view button/i)
+
+    const mapButton = canvas.getByLabelText(/map view button/i)
+    const listTab = canvas.getByTestId('list-tab')
+    await expect(listButton).toBeInTheDocument()
+    await expect(mapButton).toBeInTheDocument()
+
+    // List Tab is the default
+    await expect(listTab).toBeInTheDocument()
+
+    // Map Tap appears after clicking map button
+    await userEvent.click(mapButton)
+    await expect(listTab).toHaveAttribute('data-state', 'inactive')
+
+    // List Tab appears again after clicking list button
+    await userEvent.click(listButton)
+    await expect(listTab).toBeInTheDocument()
+  },
+}
